+++ src/types/public/permissions/domains/runningAPrison/RunningAPrisonDomainPermissions.ts
@@ -6,10 +6,15 @@ import {
   PrisonerBaseLocationPermission,
   PrisonerBaseLocationPermissions,
 } from './prisonerBaseLocation/PrisonerBaseLocationPermissions'
+import { PrisonerMovesPermission, PrisonerMovesPermissions } from './prisonerMoves/PrisonerMovesPermissions'
 
 export interface RunningAPrisonDomainPermissions {
   prisonerVisitsAndVisitors: PrisonerVisitsAndVisitorsPermissions
   prisonerBaseLocation: PrisonerBaseLocationPermissions
+  prisonerMoves: PrisonerMovesPermissions
 }
 
-export type RunningAPrisonDomainPermission = PrisonerVisitsAndVisitorsPermission | PrisonerBaseLocationPermission
+export type RunningAPrisonDomainPermission =
+  | PrisonerVisitsAndVisitorsPermission
+  | PrisonerBaseLocationPermission
+  | PrisonerMovesPermission